+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * smart_zip to be deprecated in 2.23.0
 * removed SIGINT from hang_unti_signal
 * updated ListWrapperIterator's signature
+* outfitted ListWrapperIterator with a __contains__ method